- added unit tests - added comments - pruned a little bit of code - formatting - changed naming for consistency

+++ test/jgrl/struct/se/TestSe3_F32.java
@@ -36,50 +36,50 @@ import static org.junit.Assert.assertTrue;
  */
 public class TestSe3_F32 extends GenericInvertibleTransformTests_F32<Point3D_F32> {
 
-	Random rand = new Random(234234);
+	Random rand = new Random( 234234 );
 
 	/**
 	 * Checks to see if the constructor correctly saves the reference or copies the values.
 	 */
 	@Test
 	public void constructor_assign() {
-		DenseMatrix64F R = new DenseMatrix64F(3, 3);
-		Vector3D_F32 T = new Vector3D_F32(1, 2, 3);
+		DenseMatrix64F R = new DenseMatrix64F( 3, 3 );
+		Vector3D_F32 T = new Vector3D_F32( 1, 2, 3 );
 
-		Se3_F32 a = new Se3_F32(R, T, false);
-		assertTrue(R != a.getR());
-		assertTrue(T != a.getT());
+		Se3_F32 a = new Se3_F32( R, T, false );
+		assertTrue( R != a.getR() );
+		assertTrue( T != a.getT() );
 
-		a = new Se3_F32(R, T, true);
-		assertTrue(R == a.getR());
-		assertTrue(T == a.getT());
+		a = new Se3_F32( R, T, true );
+		assertTrue( R == a.getR() );
+		assertTrue( T == a.getT() );
 	}
 
 	@Override
 	public Point3D_F32 createRandomPoint() {
-		return new Point3D_F32((float) rand.nextGaussian() * 3,
-				(float) rand.nextGaussian() * 3, (float) rand.nextGaussian() * 3);
+		return new Point3D_F32( (float)rand.nextGaussian() * 3,
+				(float)rand.nextGaussian() * 3, (float)rand.nextGaussian() * 3 );
 	}
 
 	@Override
 	public SpecialEuclidean createRandomTransform() {
 
-		float rotX = (float) ((rand.nextFloat() - 0.5f) * 2.0f * Math.PI);
-		float rotY = (float) ((rand.nextFloat() - 0.5f) * 2.0f * Math.PI);
-		float rotZ = (float) ((rand.nextFloat() - 0.5f) * 2.0f * Math.PI);
-		float x = (float) (rand.nextGaussian() * 2);
-		float y = (float) (rand.nextGaussian() * 2);
-		float z = (float) (rand.nextGaussian() * 2);
+		float rotX = (float) ( ( rand.nextFloat() - 0.5f ) * 2.0f * (float)Math.PI );
+		float rotY = (float) ( ( rand.nextFloat() - 0.5f ) * 2.0f * (float)Math.PI );
+		float rotZ = (float) ( ( rand.nextFloat() - 0.5f ) * 2.0f * (float)Math.PI );
+		float x = (float) ( (float)rand.nextGaussian() * 2 );
+		float y = (float) ( (float)rand.nextGaussian() * 2 );
+		float z = (float) ( (float)rand.nextGaussian() * 2 );
 
 		Se3_F32 ret = new Se3_F32();
 
-		SpecialEuclideanOps_F32.setEulerXYZ(rotX, rotY, rotZ, x, y, z, ret);
+		SpecialEuclideanOps_F32.setEulerXYZ( rotX, rotY, rotZ, x, y, z, ret );
 
 		return ret;
 	}
 
 	@Override
-	public Point3D_F32 apply(InvertibleTransform se, Point3D_F32 point, Point3D_F32 result) {
-		return SePointOps_F32.transform((Se3_F32) se, (Point3D_F32) point, (Point3D_F32) result);
+	public Point3D_F32 apply( InvertibleTransform se, Point3D_F32 point, Point3D_F32 result ) {
+		return SePointOps_F32.transform( (Se3_F32) se, (Point3D_F32) point, (Point3D_F32) result );
 	}
 }
